# Improve OvershadowableManager performance