--- a/core/src/main/java/org/apache/druid/timeline/partition/OvershadowableManager.java
+++ b/core/src/main/java/org/apache/druid/timeline/partition/OvershadowableManager.java
@@ -113,6 +113,22 @@ class OvershadowableManager<T extends Overshadowable<T>>
     this.overshadowedGroups = new TreeMap<>(other.overshadowedGroups);
   }
 
+  public OvershadowableManager<T> copyVisible()
+  {
+    final OvershadowableManager<T> copy = new OvershadowableManager<>();
+    visibleGroupPerRange.forEach((partitionRange, versionToGroups) -> {
+      // There should be only one group per partition range
+      final AtomicUpdateGroup<T> group = versionToGroups.values().iterator().next();
+      group.getChunks().forEach(chunk -> copy.knownPartitionChunks.put(chunk.getChunkNumber(), chunk));
+
+      copy.visibleGroupPerRange.put(
+          partitionRange,
+          new SingleEntryShort2ObjectSortedMap<>(group.getMinorVersion(), AtomicUpdateGroup.copy(group))
+      );
+    });
+    return copy;
+  }
+
   private TreeMap<RootPartitionRange, Short2ObjectSortedMap<AtomicUpdateGroup<T>>> getStateMap(State state)
   {
     switch (state) {
@@ -1088,6 +1104,12 @@ class OvershadowableManager<T extends Overshadowable<T>>
     {
       key = -1;
       val = null;
+    }
+
+    private SingleEntryShort2ObjectSortedMap(short key, V val)
+    {
+      this.key = key;
+      this.val = val;
     }
 
     @Override