--- a/core/src/main/java/org/apache/druid/timeline/partition/PartitionHolder.java
+++ b/core/src/main/java/org/apache/druid/timeline/partition/PartitionHolder.java
@@ -34,6 +34,11 @@ public class PartitionHolder<T extends Overshadowable<T>> implements Iterable<Pa
 {
   private final OvershadowableManager<T> overshadowableManager;
 
+  public static <T extends Overshadowable<T>> PartitionHolder<T> copyVisibleChunks(PartitionHolder<T> partitionHolder)
+  {
+    return new PartitionHolder<>(partitionHolder.overshadowableManager.copyVisible());
+  }
+
   public PartitionHolder(PartitionChunk<T> initialChunk)
   {
     this.overshadowableManager = new OvershadowableManager<>();
@@ -48,9 +53,14 @@ public class PartitionHolder<T extends Overshadowable<T>> implements Iterable<Pa
     }
   }
 
-  public PartitionHolder(PartitionHolder<T> partitionHolder)
+  protected PartitionHolder(OvershadowableManager<T> overshadowableManager)
   {
-    this.overshadowableManager = new OvershadowableManager<>(partitionHolder.overshadowableManager);
+    this.overshadowableManager = overshadowableManager;
+  }
+
+  public ImmutablePartitionHolder<T> asImmutable()
+  {
+    return new ImmutablePartitionHolder<>(overshadowableManager.copyVisible());
   }
 
   public boolean add(PartitionChunk<T> chunk)