--- a/server/src/main/java/org/apache/druid/client/DataSourcesSnapshot.java
+++ b/server/src/main/java/org/apache/druid/client/DataSourcesSnapshot.java
@@ -169,5 +169,4 @@ public class DataSourcesSnapshot
     }
     return overshadowedSegments;
   }
-
 }