… core/src/test/java/org/apache/druid/timeline/VersionedIntervalTimelineSpecificDataTest.java
… core/src/test/java/org/apache/druid/timeline/VersionedIntervalTimelineTestBase.java
… core/src/test/java/org/apache/druid/timeline/partition/AtomicUpdateGroupTest.java
… core/src/test/java/org/apache/druid/timeline/partition/OvershadowableManagerTest.java
… indexing-service/src/test/java/org/apache/druid/indexing/common/task/batch/parallel/SinglePhaseParallelIndexingTest.java
… integration-tests/src/test/java/org/apache/druid/tests/indexer/AbstractITBatchIndexTest.java
… server/src/test/java/org/apache/druid/server/coordinator/duty/CompactSegmentsTest.java