@suneet-s 2020-03-09
I'm not sure when to override clone and when not to. I think it's meant to be used for deepCopies, so it seems to be a good fit for the function you implemented. Only change that probably matters is in NewestSegmentFirstIterator